chore: #6 HomeActivity -> MainActivity 수정
