+++ app/src/main/AndroidManifest.xml
@@ -35,12 +35,6 @@
             android:label="@string/app_name"
             android:theme="@style/Theme.NOWSOPTAndroid">
         </activity>
-        <activity
-            android:name=".HomeActivity"
-            android:exported="false"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.NOWSOPTAndroid">
-        </activity>
     </application>
 
 </manifest>
